Partition tests by rules
